requirements.txt
autopep8==1.6.0
cffi==1.15.0
chardet==4.0.0
click==8.0.3
cryptography==36.0.0
jellyfish==0.8.9
networkx==2.6.3
numpy==1.21.4
pdfminer.six==20211012
pycodestyle==2.8.0
pycparser==2.21
regex==2021.11.10
segtok==1.5.10
tabulate==0.8.9
toml==0.10.2
yake==0.4.8
